+++ src/front/components/GlobalRankingWidget.jsx
@@ -1,5 +1,3 @@
-// src/front/components/GlobalRankingWidget.jsx
-
 import React from "react";
 import { Link } from "react-router-dom";
 import Avatars from "./Avatars";
@@ -9,14 +7,12 @@ export function GlobalRankingWidget({ entries }) {
     <div
       className="p-4 rounded-lg mb-6"
       style={{
-        backgroundColor: "var(--color-info)",   // azul destacado
-        color: "var(--color-white)",            // texto blanco
+        backgroundColor: "var(--color-info)",
+        color: "var(--color-white)",
       }}
     >
       {/* Título */}
-      <h3 className="text-lg font-semibold mb-4">
-        Global Ranking
-      </h3>
+      <h3 className="text-lg font-semibold mb-4">Global Ranking</h3>
 
       {/* Tres primeros puestos */}
       <Avatars entries={entries} />
@@ -27,7 +23,7 @@ export function GlobalRankingWidget({ entries }) {
           <button
             className="px-4 py-2 rounded-lg font-medium"
             style={{
-              backgroundColor: "var(--color-accent)",  // rosa accent
+              backgroundColor: "var(--color-accent)",
               color: "var(--color-white)",
             }}
           >
